+++ src/components/sodaforsparc/hero.tsx
@@ -2,14 +2,14 @@ import { useState, useEffect } from 'react';
 
 import Image from 'next/image';
 
-import getLatestURL from '@/lib/getLatestDownloadLink';
+import getLatestDownloadLink from '@/lib/getLatestDownloadLink';
 
 export default function Hero() {
   const [downloadURL, setDownloadURL] = useState<string | undefined>(``);
 
   useEffect(() => {
     const func = async () => {
-      const url = await getLatestURL(`fairdataihub/SODA-for-SPARC`);
+      const url = await getLatestDownloadLink(`fairdataihub/SODA-for-SPARC`);
       setDownloadURL(url);
     };
     func();
@@ -52,7 +52,7 @@ export default function Hero() {
                 <div className="flex flex-row justify-center">
                   <a
                     href={downloadURL}
-                    className="rounded border-0 border-none bg-black px-6 py-2 text-lg text-white ring-2 ring-transparent ring-offset-2 transition-all hover:ring-pink-600 focus:outline-none focus:ring-pink-600 sm:block"
+                    className="umami--click--soda-sparc-download-button rounded border-0 border-none bg-black px-6 py-2 text-lg text-white ring-2 ring-transparent ring-offset-2 transition-all hover:ring-pink-600 focus:outline-none focus:ring-pink-600 sm:block"
                   >
                     Download now
                   </a>
@@ -62,7 +62,7 @@ export default function Hero() {
                 href="https://docs.sodaforsparc.io/"
                 target="_blank"
                 rel="noreferrer"
-                className="flex flex-row justify-center"
+                className="umami--click--soda-sparc-docs-button flex flex-row justify-center"
               >
                 <button className=" rounded border-none bg-black px-6 py-2 text-lg text-white ring-2 ring-transparent ring-offset-2 transition-all hover:ring-pink-600 focus:outline-none focus:ring-pink-600">
                   Explore the docs