+++ src/components/sparclink/info.tsx
@@ -11,8 +11,8 @@ export default function Info() {
     },
     {
       name: `Sachira Kuruppu`,
-      href: `/sparclink`,
-      external: false,
+      href: `https://www.linkedin.com/in/sachirakuruppu/`,
+      external: true,
       image: `https://avatars.dicebear.com/api/jdenticon/SachiraKuruppu.svg`,
     },
     {
@@ -58,7 +58,7 @@ export default function Info() {
                     <a
                       href="https://github.com/fairdataihub/SPARClink/graphs/contributors"
                       target="_blank"
-                      className="mr-2"
+                      className="umami--click--sparclink-contributors-badge mr-2"
                       rel="noreferrer"
                     >
                       {/*  eslint-disable-next-line @next/next/no-img-element */}
@@ -70,7 +70,7 @@ export default function Info() {
                     <a
                       href="https://github.com/fairdataihub/SPARClink/stargazers"
                       target="_blank"
-                      className="mr-2"
+                      className="umami--click--sparclink-stars-badge mr-2"
                       rel="noreferrer"
                     >
                       {/*  eslint-disable-next-line @next/next/no-img-element */}
@@ -82,7 +82,7 @@ export default function Info() {
                     <a
                       href="https://github.com/fairdataihub/SPARClink/issues"
                       target="_blank"
-                      className="mr-2"
+                      className="umami--click--sparclink-issues-badge mr-2"
                       rel="noreferrer"
                     >
                       {/*  eslint-disable-next-line @next/next/no-img-element */}
@@ -94,7 +94,7 @@ export default function Info() {
                     <a
                       href="https://github.com/fairdataihub/SPARClink/blob/master/LICENSE"
                       target="_blank"
-                      className="mr-2"
+                      className="umami--click--sparclink-license-badge mr-2"
                       rel="noreferrer"
                     >
                       {/*  eslint-disable-next-line @next/next/no-img-element */}
@@ -108,7 +108,7 @@ export default function Info() {
                     <a
                       href="https://github.com/fairdataihub/SPARClink"
                       target="_blank"
-                      className="text-url hover-underline-animation"
+                      className="text-url hover-underline-animation umami--click--sparclink-github"
                       rel="noreferrer"
                     >
                       <span className="font-lato">
@@ -134,6 +134,7 @@ export default function Info() {
                     target="_blank"
                     rel="noreferrer"
                     aria-label="Github"
+                    className="umami--click--sparclink-github"
                   >
                     <svg
                       xmlns="http://www.w3.org/2000/svg"
@@ -172,7 +173,7 @@ export default function Info() {
                     <a
                       href="https://sparc.science/help/2021-sparc-fair-codeathon"
                       target="_blank"
-                      className="text-url hover-underline-animation"
+                      className="text-url hover-underline-animation umami--click--sparc-fair-21-codeathon"
                       rel="noreferrer"
                     >
                       <span className="font-lato">